pyproject.toml
[tool.poetry]
name = "rtfs"
version = "0.1.1"
description = "Code Repo Summary Generator"
authors = ["John Peng <kongyijipeng@gmail.com>"]
readme = "README.md"

packages = [
    { include = "rtfs" },
]
include = [
    "rtfs/languages/**/*.so",
    "rtfs/languages/**/*.json",
    "rtfs/languages/**/*.scm",
    "rtfs/moatless/**/*.scm",
]

[tool.poetry.dependencies]
python = "^3.8"
anthropic = "0.32.0"
infomap = "2.8.0"
intervaltree = "3.1.0"
llama_index = "0.10.59"
networkx = "3.3"
openai = "1.38.0"
pydantic = "2.8.2"
python-dotenv = "1.0.1"
PyYAML = "6.0.1"
setuptools = "70.1.1"
simple_parsing = "0.1.5"
starlette = "0.38.2"
tenacity = "8.5.0"
tiktoken = "0.7.0"
tree_sitter = "0.22.3"
tree_sitter_java = "0.21.0"
tree_sitter_python = "0.21.0"
typing_extensions = "4.12.2"

[tool.poetry.scripts]
rtfs = "rtfs.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
